requirements.txt
asgiref==3.5.2
dj-database-url==1.0.0
Django==4.1.1
django-crispy-forms==1.14.0
django-user-agents==0.4.0
gunicorn==20.1.0
psycopg2-binary==2.9.3
sqlparse==0.4.2
tzdata==2022.2
ua-parser==0.16.1
user-agents==2.2.0
whitenoise==6.2.0
